requirements.txt
beautifulsoup4==4.13.4
google_api_python_client==2.170.0
google_auth_oauthlib==1.2.2
protobuf==6.31.0
python-dotenv==1.1.0
Requests==2.32.3
